[HQ-3390] want to ignore removing proto types that do not exist since there are no associated resources
Former-commit-id: c312fe70c1f4abc63beef5e4af98e49bf1c0a1b5

+++ hq-server/src/main/java/org/hyperic/hq/authz/server/session/ResourceManagerImpl.java
@@ -1430,10 +1430,16 @@ public class ResourceManagerImpl implements ResourceManager, ApplicationContextA
         for (final String typeName : typeNames) {
             removeMeasurementTempls(typeName);
             final MonitorableType mt = monitorableTypeDAO.findByName(typeName);
-            monitorableTypeDAO.remove(mt);
+            // mt should not be null at this point, but you never know
+            if (mt != null) {
+                monitorableTypeDAO.remove(mt);
+            }
             final Resource proto = resourceDAO.findResourcePrototypeByName(typeName);
-            removeAppdefType(proto);
-            resourceDAO.remove(proto);
+            // proto = null here means that no resources are configured with that particular type
+            if (proto != null) {
+                removeAppdefType(proto);
+                resourceDAO.remove(proto);
+            }
         }
     }
